feat(diff): show more usefull panes
Tab1 is a two-way diff of just the conflicts.
Resolve the conflicts here (in the left window) and save the file.
    +-----------------------------------------------------------------+
    |  LOCAL Conflicts (solve here)  |        REMOTE conflicts        |
    +-----------------------------------------------------------------+

Tab2 is a two-way diff of the REMOTE file (incoming changes) and the merge base (previous known version).
    +-----------------------------------------------------------------+
    |          REMOTE file           |           BASE file            |
    +-----------------------------------------------------------------+

Tab3 is a two-way diff of the LOCAL file (your changes) and the merge base (previous known version).
    +-----------------------------------------------------------------+
    |           LOCAL file           |           BASE file            |
    +-----------------------------------------------------------------+